+++ TidalPy/tides/multilayer/numerical_int/__init__.py
@@ -1,3 +1,4 @@
 from .derivatives import static_solid_ode, static_liquid_ode, dynamic_liquid_ode, dynamic_solid_ode
 from .initial_conditions import known_initial_guess_funcs, find_initial_guess
 from .solver import tidal_y_solver
+from .solver_numba import tidal_y_solver as tidal_y_solver_numba